OpenMonitor/django_app/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
requests
python-dotenv
numpy